Use intp to match standard ints

+++ tests/epyccel/test_functionals.py
@@ -1,6 +1,6 @@
 # pylint: disable=missing-function-docstring, missing-module-docstring
 from numpy.random import randint
-from numpy import equal
+from numpy import equal, intp
 import pytest
 
 
@@ -20,11 +20,11 @@ def test_functional_for_overwrite_1d_range(language):
     compare_epyccel(functionals.functional_for_overwrite_1d_range, language)
 
 def test_functional_for_1d_var(language):
-    y = randint(99,size = 4)
+    y = randint(99, size = 4, dtype = np.intp)
     compare_epyccel(functionals.functional_for_1d_var, language, y)
 
 def test_functional_for_1d_const(language):
-    y = randint(99,size = 4)
+    y = randint(99, size = 4, dtype = np.intp)
     z = randint(99)
     compare_epyccel(functionals.functional_for_1d_const, language, y, z)
 
@@ -35,12 +35,12 @@ def test_functional_for_2d_range(language):
     compare_epyccel(functionals.functional_for_2d_range, language)
 
 def test_functional_for_2d_var_range(language):
-    y = randint(99, size = 3)
+    y = randint(99, size = 3, dtype = np.intp)
     compare_epyccel(functionals.functional_for_2d_var_range, language, y)
 
 def test_functional_for_2d_var_var(language):
-    y = randint(99, size = 3)
-    z = randint(99, size = 2)
+    y = randint(99, size = 3, dtype = np.intp)
+    z = randint(99, size = 2, dtype = np.intp)
     compare_epyccel(functionals.functional_for_2d_var_var, language, y, z)
 
 def test_functional_for_2d_dependant_range(language):
@@ -71,5 +71,5 @@ def test_functional_for_3d_range(language):
     compare_epyccel(functionals.functional_for_3d_range, language)
 
 def test_unknown_length_functional(language):
-    y = randint(100, size = 20)
+    y = randint(100, size = 20, dtype = np.intp)
     compare_epyccel(functionals.unknown_length_functional, language, y)
